Some minor format string fixes for gcc-3.3.3 (SuSE 9.1).  Not sure why later gccs don't complain about these.
git-svn-id: svn://svn.valgrind.org/valgrind/trunk@9019

+++ coregrind/m_syswrap/syswrap-generic.c
@@ -2029,7 +2029,7 @@ ML_(generic_PRE_sys_mmap) ( ThreadId tid,
 #define POST(name)     DEFN_POST_TEMPLATE(generic, name)
 
 // Combine two 32-bit values into a 64-bit value
-#define LOHI64(lo,hi)   ( (lo) | ((ULong)(hi) << 32) )
+#define LOHI64(lo,hi)   ( ((ULong)(lo)) | (((ULong)(hi)) << 32) )
 
 //zz //PRE(sys_exit_group, Special)
 //zz //{
